refactor: change package locations
